requirements/base.txt
Django==1.10

# to read environment variables from .env file
django-dotenv==1.4.1
